requirements.txt
flask
openai-whisper
torch
google-generativeai
ffmpeg-python
